+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* `flexberry-groupedit` component:
+    * Now possible specify `developerUserSettings` for `flexberry-groupedit` on `edit-form`.
 
 ## [0.6.2-beta.1] - 2016-09-23
 ### Fixed